functions/sample/python/requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.4
ibm-cloud-sdk-core==3.10.0
ibmcloudant==0.0.34
idna==3.2
jproperties==2.1.1
PyJWT==2.1.0
python-dateutil==2.8.2
requests==2.26.0
six==1.16.0
urllib3==1.26.6
websocket-client==1.1.0
# cloudant
